+++ TutorialAppKotlin/app/src/main/kotlin/com/generalmagic/gemsdk/demo/activities/mainactivity/MainActivity.kt
@@ -12,6 +12,7 @@ package com.generalmagic.gemsdk.demo.activities.mainactivity
 
 import android.Manifest
 import android.content.Context
+import android.graphics.Color
 import android.os.Build
 import android.os.Bundle
 import android.view.View
@@ -125,7 +126,16 @@ class MainActivity : BaseActivity(), IMapControllerActivity {
         setSupportActionBar(toolbar)
 
         // make status bar transparent
-        window.addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
+        {
+            val c: Int = window.statusBarColor
+            window.statusBarColor = Color.argb(128, Color.red(c), Color.green(c), Color.blue(c))
+            window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_STABLE or View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
+        }
+        else
+        {
+            window.setFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS, WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
+        }
 
         // no title
         supportActionBar?.title = ""
@@ -240,7 +250,8 @@ class MainActivity : BaseActivity(), IMapControllerActivity {
             Manifest.permission.READ_PHONE_STATE,
             Manifest.permission.READ_CONTACTS,
             Manifest.permission.WRITE_CONTACTS,
-            Manifest.permission.BLUETOOTH
+            Manifest.permission.BLUETOOTH,
+            Manifest.permission.MODIFY_AUDIO_SETTINGS
         )
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {