+++ TutorialAppKotlin/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2020-12-23
+### Added
+- UploadLog form activity.
+- LogUploader instance kept in GemApplication.
+
+### Fixed
+- not declared MODIFY_AUDIO_SETTINGS permission.
+
 ## [2.5.0] - 2020-12-16
 ### Added
 - Audio setting and used it in log recorder tutorial.